0.1.42: say what is actually wrong when an import cannot see its files
A download client can be configured to report host paths rather than the paths it
writes inside its own container. Bookmarkarr stored that string verbatim, failed to
open it, and reported "No importable files found" — true, and useless, because it
points at the files when the directory was never visible. The client's Test passed
too, since it only exercised the API.

Imports now distinguish nothing-reported, every-directory-unreachable (named as a
remote path mapping problem), and partially-reachable. Test samples the reported
paths and says the same thing, keeping Success true because the connection itself
is fine.

Two silences behind that: an id-filtered poll returning none of the downloads it
asked about logged nothing and left every row frozen, and nothing tracked how long
a download had gone without changing. Both are now reported.

Also fixed:

- The update endpoint still returned the tracked entity, so Audiobook -> Editions ->
  Book -> Editions recursed past the serializer depth limit and the caller saw a 500
  for a change that had already committed.
- Audiobook.BasePath normalised on read, and normalising a blank string resolves it
  against the process working directory, so "no base path" came back as a real path
  pointing wherever the app was running — including to the guards meant to reject it.
- The download client dialog offered a remote path mapping picker writing a
  remotePathMappingIds list no server code has ever read, under help text promising
  the opposite of what happens. The server resolves by the mapping's own
  downloadClientId; the UI now matches.
- Size limits above 2047 MB overflowed to a negative number and rejected every
  release. Found while adding the 5 GB default below, which would have triggered it
  on every install.

Added a 5 GB default MaximumSize on new quality profiles — a guard against a
mislabelled release that is a whole series, not a quality judgement. Seeding stays
additive, so existing profiles keep whatever their owner chose.

Tooltips are now anchored to their control rather than drawn at the pointer, where
an enlarged cursor covers the first line and no styling can move it.

+++ bookmarkarr.application/Search/Scoring/SearchResultScorer.cs
@@ -92,13 +92,17 @@ namespace Bookmarkarr.Application.Search.Scoring
             // Size checks (skip for NZB)
             if (!isNzb && searchResult.Size > 0)
             {
-                if (profile.MinimumSize > 0 && searchResult.Size < profile.MinimumSize * 1024 * 1024)
+                // The casts are load-bearing. MinimumSize/MaximumSize are megabytes held in an int,
+                // so the byte conversion overflowed for any limit above 2047 MB and wrapped to a
+                // negative number — at which point every release is "larger than the maximum" and
+                // the profile silently rejects everything it is asked to score.
+                if (profile.MinimumSize > 0 && searchResult.Size < (long)profile.MinimumSize * 1024 * 1024)
                 {
                     score.RejectionReasons.Add($"File too small (< {profile.MinimumSize} MB)");
                     score.TotalScore = -1;
                     return score;
                 }
-                if (profile.MaximumSize > 0 && searchResult.Size > profile.MaximumSize * 1024 * 1024)
+                if (profile.MaximumSize > 0 && searchResult.Size > (long)profile.MaximumSize * 1024 * 1024)
                 {
                     score.RejectionReasons.Add($"File too large (> {profile.MaximumSize} MB)");
                     score.TotalScore = -1;
